requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.7.8
bs4==0.0.1
catalogue==2.0.8
certifi==2022.6.15
chardet==4.0.0
click==8.0.4
cloudpickle==2.1.0
colorama==0.4.5
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
executing==1.0.0
fake-useragent==0.1.11
fonttools==4.37.1
fsspec==2022.8.0
future==0.18.2
google==3.0.0
googlesearch-python==1.0.1
graphviz==0.20.1
icecream==2.1.3
idna==2.10
importlib-metadata==4.12.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
langcodes==3.3.0
locket==1.0.0
lxml==4.6.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
murmurhash==1.0.8
nest-asyncio==1.5.5
networkx==2.8.6
nltk==3.7
numpy==1.21.0
packaging==21.3
pandas==1.4.4
parse==1.19.0
parso==0.8.3
partd==1.3.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.7
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.8.2
pyee==8.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pytextrank==3.2.4
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.25.1
requests-html==0.10.0
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.0
thinc==8.0.17
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
typing-extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.12
w3lib==2.0.1
wasabi==0.10.1
wcwidth==0.2.5
websockets==10.3
zipp==3.8.1
